Allow player choice in responding to forbidden challenge
Previously, the bot would automatically select to flip all cards if
player did not have, or flip the forbidden card if player had.

However, if player has 2 influence yet does not have the forbidden
character, it is still a valid option to flip a character anyway!
This means we need to allow a "!flip all" command.